+++ kibitzr/fetcher/browser/launcher.py
@@ -66,8 +66,6 @@ def firefox(headless=True):
             firefox_binary=firefox_binary,
             firefox_profile=firefox_profile,
         )
-        FIREFOX_INSTANCE[driver_key].set_window_size(1024, 768)
-        FIREFOX_INSTANCE[driver_key].implicitly_wait(2)
     yield FIREFOX_INSTANCE[driver_key]