+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
 ## Unreleased
+
+## [4.0.4] - 2017-06-07
 ## Changed
 - Firefox fetcher: Implicitly wait 2 seconds for selects
 - Firefox fetcher: Resize window before each fetch
+## Fixed
+- bash transform: Skip execution for empty content
 
 ## [4.0.3] - 2017-05-25
 ## Added